client/requirements.txt
wxPython==4.0.7.post2
requests==2.23.0
pycryptodome==3.9.7
